pyproject.toml
[tool.poetry]
name = "to-do-list"
version = "0.1.0"
description = ""
authors = ["NickolasSilva01 <nickolas.silva@ottech.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.114.2"}
ignr = "^2.2"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.12"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev to_do_list/app.py --port 8003'
pre_test = 'task lint'
test = 'pytest --cov=to_do_list -vv'
post_test = 'coverage html'
lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . --fix ; ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
